+++ history.md
@@ -4,7 +4,7 @@
 #### v0.1.22
 - Added method `mpmToolbox.projectData.alignment.Alignment.repositionAll()`.
 - Enhancement to method `mpmToolbox.gui.MpmToolbox.closeProject()`, when the project is closed or a new project is loaded while playback is running, the playback will now stop properly.
-- Enhanced class `mpmToolbox.projectData.alignment.AbstractAlignmentComputation` and updated class `mpmToolbox.gui.audio.AudioDocumentData` accordingly. The latter has two additional buttons to select an alignment algorithm and execute it.
+- Enhanced class `mpmToolbox.projectData.alignment.AbstractAlignmentComputation` and updated class `mpmToolbox.gui.audio.AudioDocumentData` accordingly. The latter has two additional buttons to select an alignment algorithm and execute it. Also class `mpmToolbox.projectData.alignment.PlaceholderAligner` has been added to illustrate the usage of the abstract class.
 
 
 #### v0.1.21